requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
youtube-transcript-api>=0.6.1